Refactor item details to force logic
By directly adding details certain checks were skipped that could cause problems in other processes. Now the only way to add details is by calling the correct method.
